Mocked sendDataJson and checked caching ability
